split MatrixConversions
